restore dialog stories
